refactor: rename model filtering trait
