+++ tests/Models/User.php
@@ -3,14 +3,14 @@
 namespace Kettasoft\Filterable\Tests\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use Kettasoft\Filterable\Traits\HasFilterable;
+use Kettasoft\Filterable\Traits\InteractsWithFilterable;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Kettasoft\Filterable\Tests\Database\Factories\UserFactory;
 
 class User extends Model
 {
-    use HasFactory, HasFilterable;
+    use HasFactory, InteractsWithFilterable;
 
     protected $fillable = ['name', 'email', 'is_blocked', 'platform', 'password'];
