+++ tests/Unit/Filterable/FilterableModelAutoBindingTest.php
@@ -5,6 +5,7 @@ namespace Kettasoft\Filterable\Tests\Unit\Filterable;
 use Illuminate\Database\Eloquent\Model;
 use Kettasoft\Filterable\Tests\TestCase;
 use Kettasoft\Filterable\Traits\HasFilterable;
+use Kettasoft\Filterable\Traits\InteractsWithFilterable;
 use Illuminate\Contracts\Database\Query\Builder;
 use Kettasoft\Filterable\Tests\Http\Filters\PostFilter;
 use Kettasoft\Filterable\Exceptions\FilterClassNotResolvedException;
@@ -14,7 +15,7 @@ class FilterableModelAutoBindingTest extends TestCase
   public function test_it_applies_filter_automatically_from_model_property()
   {
     $model = new class extends Model {
-      use HasFilterable;
+      use InteractsWithFilterable;
       protected $filterable = PostFilter::class;
     };
 
@@ -24,11 +25,21 @@ class FilterableModelAutoBindingTest extends TestCase
   public function test_it_throws_exception_if_no_filter_class_and_no_model_property()
   {
     $model = new class extends Model {
-      use HasFilterable;
+      use InteractsWithFilterable;
     };
 
     $this->expectException(FilterClassNotResolvedException::class);
 
     $model->filter();
+  }
+
+  public function test_legacy_trait_remains_compatible()
+  {
+    $model = new class extends Model {
+      use HasFilterable;
+      protected $filterable = PostFilter::class;
+    };
+
+    $this->assertInstanceOf(Builder::class, $model->filter());
   }
 }
